Updated `production.js` link and value
The link to `production.js` is not from a file that is actually included in the tutorial. The file that was previously linked had the value `STATIC_FILES` instead of `STATIC_FILES_OFF`, which will be the case for someone actually following the tutorial.

I also changed the `heroku config:set ...` line to match the actual value.